Version 1.1/files/requirements.txt
cryptography;
pillow;

